requirements.txt
Flask
deepgram-sdk